+++ Q3E/src/main/jni/darkmod/sys/stub/sys_stub.cpp
@@ -33,12 +33,12 @@ int Sys_Milliseconds( void ) {
 	return frameNum * 16;
 }
 
-double Sys_GetClockTicks( void ) {
-	return frameNum * 16.0;
+uint64 Sys_GetClockTicks( void ) {
+	return frameNum * 16;
 }
 
-double Sys_ClockTicksPerSecond( void ) {
-	return 1000.0;
+uint64 Sys_ClockTicksPerSecond( void ) {
+	return 1000;
 }
 
 void	Sys_Sleep( int msec ) {
